bugfixes to casting and guessing

+++ muon_detector.h
@@ -13,7 +13,7 @@ class MuonDetector : public SubDetector
 private:
   enum tracker_layers{INNER_MUON_LAYER=0, OUTER_MUON_LAYER=1};
   std::unique_ptr<vector<bool>> hit_layers; // Unique pointer to a vector of booleans.
-  int percent_chance;
+  double percent_chance;
 public:
   // Constructor
   // Default
